% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProb.R
\name{getProb}
\alias{getProb}
\title{Get the probabilities}
\usage{
getProb(log.rr, log.gop)
}
\arguments{
\item{log.rr}{A vetor of logarithm of relative risks.}

\item{log.gop}{A number oflogarithm of generalized odds product.}
}
\value{
A vector of probabilities
}
\description{
By having logarithm of relative risks and logarithm of generalized odds product, the probability of event from each each treatment is calculated.
}
\examples{
getProb(c(1, 10, 5), 2)

}
