% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem.mono.R
\name{mem.mono}
\alias{mem.mono}
\title{Multiplicative effect modeling with monotonic treatment effect assumption}
\usage{
mem.mono(y, z, va, vb, alpha.start, beta.start, max.step = 10^4,
  thres = 10^-6)
}
\arguments{
\item{y}{The response vector. Should only take values 0 or 1.}

\item{z}{The treatment vector. Should be numerical.}

\item{va}{The covariates matrix for the target model.}

\item{vb}{The covariates matrix for the nuisance model.}

\item{alpha.start}{Starting values for the parameters in the target model.}

\item{beta.start}{Starting values for the parameters in the nuisance model.}

\item{max.step}{The maximal number of iterations to be passed into the optim function. The default is 10,000.}

\item{thres}{Threshold for judging convergence. The default is 1e-6.}
}
\value{
A list consisting of
 \itemize{
   \item point.est - the point estimates.
   \item se.est - he standard error estimates.
   \item cov - estimate of the covariance matrix for the estimates.
   \item conf.lower - the lower limit of the 95\% confidence interval.
   \item conf.upper - the upper limit of the 95\% confidence interval.
   \item p.value - the two sided p-value for testing zero coefficients.
   \item coefficients - the matrix summarizing key information: point estimate, 95\% confidence interval and p-value.
   \item param - the measure of association.
   \item converged - Did the maximization process converge?
   \item likelihood - The likelihood under the estimates.
}
}
\description{
\code{mem.mono} is used to estimate the multiplicative effect under monotonic treatment effect assumption. Treatment can be continuous or ordinal.
}
\details{
The estimation is based on the models: \eqn{log(RR(z0,z;v,\alpha)) = \alpha*v(z-z0)},
\eqn{log(RR(z_min,z_max;v,\beta)) = \beta*v}, where \eqn{RR(z0,z) = P(Y=1|Z=z,z)/(Y=1|Z=z,z0)}.
}
